Moved seal methods in dedicated class and declaring deprecated old methods
